rename src/app
