+++ src/app/mysterion.js
@@ -28,7 +28,7 @@ import { onFirstEvent, onFirstEventOrTimeout } from './communication/utils'
 import path from 'path'
 import ConnectionStatusEnum from '../libraries/mysterium-tequilapi/dto/connection-status-enum'
 import type { Size } from './window'
-import type { MysterionConfig } from './mysterionConfig'
+import type { MysterionConfig } from './mysterion-config'
 import Window from './window'
 import Terms from './terms'
 import ProcessMonitoring from '../libraries/mysterium-client/monitoring'
