+++ src/renderer/dependencies/modules/application.js
@@ -25,7 +25,7 @@ import NullCollector from '../../../app/statistics/null-collector'
 import { remote } from 'electron'
 import type { ApplicationInfo } from '../../../app/statistics/events'
 import { createEventFactory } from '../../../app/statistics/events'
-import VpnInitializer from '../../../app/vpnInitializer'
+import VpnInitializer from '../../../app/vpn-initializer'
 import type { TequilapiClient } from '../../../libraries/mysterium-tequilapi/client'
 import realSleep from '../../../libraries/sleep'
 import IpcMessageBus from '../../../app/communication/ipc-message-bus'
