+++ test/unit/app/vpnInitializer.spec.js
@@ -17,7 +17,7 @@
 
 // @flow
 
-import VpnInitializer from '../../../src/app/vpnInitializer'
+import VpnInitializer from '../../../src/app/vpn-initializer'
 import { beforeEach, describe, expect, it } from '../../helpers/dependencies'
 import type { State as IdentityState } from '../../../src/renderer/store/modules/identity'
 import IdentityDTO from '../../../src/libraries/mysterium-tequilapi/dto/identity'
